@@ -4,7 +4,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.ice</groupId>
 	<artifactId>jcvs</artifactId>
-	<version>5.7.0-SNAPSHOT</version>
+	<version>5.8.0-SNAPSHOT</version>
 
 	<licenses>
 		<license>
@@ -32,8 +32,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.1</version>
 				<configuration>
-					<source>1.7</source>
-					<target>1.7</target>
+					<source>1.8</source>
+					<target>1.8</target>
 					<fork>false</fork>
 					<showDeprecation>true</showDeprecation>
 					<showWarnings>true</showWarnings>
