@@ -120,12 +120,7 @@ extends		Thread
 					( new InputStreamReader
 						( this.proc.getErrorStream() ) );
 
-			final Thread t = new Thread(
-				new Runnable()
-					{
-					@Override
-					public void
-					run()
+			final Thread t = new Thread(() ->
 						{
 						try {
 							for ( ; ; )
@@ -143,7 +138,6 @@ extends		Thread
 							appendEx( ex, "reading exec stderr stream" );
 							}
 						}
-					}
 				);
 
 			t.start();
